fix(packages/wagmi): move common-bases to pinnedTokens hook
