+++ apps/docs/.storybook/preview.ts
@@ -1,23 +1,26 @@
 import '@sushiswap/ui/src/index.css'
 import './style.css'
-
-import { withThemeByClassName } from '@storybook/addon-styling'
 import { Preview } from '@storybook/react'
-
-export const decorators = [
-  withThemeByClassName({
-    themes: {
-      light: 'light',
-      dark: 'dark',
-    },
-    defaultTheme: 'light',
-  }),
-]
 
 export const preview: Preview = {
   parameters: {
     backgrounds: {
       disable: true,
+    },
+  },
+  globalTypes: {
+    theme: {
+      description: 'Global theme for components',
+      defaultValue: 'light',
+      toolbar: {
+        // The label to show for this toolbar item
+        title: 'Theme',
+        icon: 'circlehollow',
+        // Array of plain string values or MenuItem shape (see below)
+        items: ['light', 'dark'],
+        // Change title based on selected value
+        dynamicTitle: true,
+      },
     },
   },
 }
