+++ apps/evm/ui/swap/search/SearchPanel.tsx
@@ -2,19 +2,15 @@
 
 import { ExclamationTriangleIcon } from '@heroicons/react/24/outline'
 import { Chain } from '@sushiswap/chain'
-import { Token } from '@sushiswap/currency'
-import { useDebounce } from '@sushiswap/hooks'
-import { usePrice, useTokenList, useTokenSearch } from '@sushiswap/react-query'
-import { COMMON_BASES } from '@sushiswap/router-config'
-import { classNames } from '@sushiswap/ui'
-import { Badge } from '@sushiswap/ui/components/Badge'
+import { Native, Token } from '@sushiswap/currency'
+import { useCustomTokens, useDebounce, usePinnedTokens } from '@sushiswap/hooks'
+import { usePrice, useTokenList, useTokens, useTokenSearch } from '@sushiswap/react-query'
+import { Badge, classNames, NetworkIcon, SkeletonCircle, SkeletonText } from '@sushiswap/ui'
 import { Button } from '@sushiswap/ui/components/button'
 import { Currency } from '@sushiswap/ui/components/currency'
 import { Dialog } from '@sushiswap/ui/components/dialog'
-import { NetworkIcon } from '@sushiswap/ui/components/icons'
 import { Search } from '@sushiswap/ui/components/input/Search'
 import { List } from '@sushiswap/ui/components/list/List'
-import { SkeletonCircle, SkeletonText } from '@sushiswap/ui/components/skeleton'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@sushiswap/ui/components/tooltip'
 import React, { FC, useCallback, useMemo, useState } from 'react'
 
@@ -29,9 +25,23 @@ export const SearchPanel: FC = () => {
 
   const { data: tokenSearch } = useTokenSearch({ address: debouncedQuery })
   const { data: tokenList } = useTokenList(filter)
+  const { data: pinnedTokenMap } = usePinnedTokens()
+  const { data: tokenMap } = useTokens({ chainId: network1 })
+  const { data: customTokenMap } = useCustomTokens()
+
+  const pinnedTokens = useMemo(() => {
+    return pinnedTokenMap[network1]
+      .map((id) => {
+        const [, address] = id.split(':')
+        if (address === 'NATIVE') return Native.onChain(network1)
+        return tokenMap?.[address] || customTokenMap?.[address]
+      })
+      .filter((token): token is Token => !!token)
+  }, [customTokenMap, network1, pinnedTokenMap, tokenMap])
+
   const { data: popularTokensList, isLoading: isPopularListLoading } = useTokenList(
     // TODO: we should have ETH as an option...
-    useMemo(() => [...COMMON_BASES[network1].map((t) => t.wrapped.address)], [network1])
+    useMemo(() => [...pinnedTokens.map((t) => t.wrapped.address)], [pinnedTokens])
   )
   const { open, setOpen } = useSearchContext()
