+++ config/router/src/index.ts
@@ -1,4 +1,3 @@
 export { ADDITIONAL_BASES } from './additional-bases.js'
 export { BASES_TO_CHECK_TRADES_AGAINST } from './bases-to-check-trades-against.js'
-export { COMMON_BASES } from './common-bases.js'
 export { CUSTOM_BASES } from './custom-bases.js'
